利用ajax上传文件，增加上传进度条，不强制上传图片，增加伪造图片扩展名验证，增加后台文件夹名称，增加favicon.ico

+++ heat.py
@@ -2,15 +2,15 @@
 # -*- coding:UTF-8-sig -*-
 
 # 虚拟环境需要的包：
-# flask, flask_uploads, Pillow
+# flask, flask_uploads(非原版), Pillow
 
 import os
 import io
 import zipfile
 from shutil import copyfile
 import csv
-from flask import Flask, flash, request, redirect, url_for, render_template, escape, send_file, session
-from flask_uploads import UploadSet, configure_uploads, IMAGES, patch_request_class, UploadNotAllowed
+from flask import Flask, flash, request, redirect, url_for, render_template, escape, send_file, jsonify
+from flask_uploads import UploadSet, configure_uploads, IMAGES, patch_request_class
 from datetime import datetime, timedelta, timezone
 from PIL import Image, ImageDraw, ImageFont
 
@@ -27,22 +27,26 @@ def seekgps(filepath, targetcol, offset, data):
     pass
 
 def scale_tag(old_file, new_path, size, text):
-    with Image.open(old_file) as im:
-        (ow, oh) = im.size
-        scale = size/max(ow, oh)
-        out = im.resize((int(ow*scale), int(oh*scale)))
+    try:
+        with Image.open(old_file) as im:
+            (ow, oh) = im.size
+            scale = size/max(ow, oh)
+            out = im.resize((int(ow*scale), int(oh*scale)))
 
-        # 给图片加标签
-        fnt = ImageFont.truetype('static/fonts/SourceHanSansSC-Medium.otf', 36)
-        d = ImageDraw.Draw(out)
-        d.polygon([(30,60), (40,50), (460,50), (445,78), (460,105), (40,105), (30,95)], fill=(4,188,212))
-        d.text((45,50), text, font=fnt, fill=(255,255,255))
+            # 给图片加标签
+            fnt = ImageFont.truetype('static/fonts/SourceHanSansSC-Medium.otf', 36)
+            d = ImageDraw.Draw(out)
+            d.polygon([(30,60), (40,50), (460,50), (445,78), (460,105), (40,105), (30,95)], fill=(4,188,212))
+            d.text((45,50), text, font=fnt, fill=(255,255,255))
 
-        if not os.path.exists(new_path):
-            os.makedirs(new_path)
-        new_file = new_path + '/' + os.path.split(im.filename)[1]
-        out.save(new_file)
-    return new_file
+            if not os.path.exists(new_path):
+                os.makedirs(new_path)
+            new_file = new_path + '/' + os.path.split(im.filename)[1]
+            out.save(new_file)
+    except Image.UnidentifiedImageError:
+        return 'fake_img'
+    else:
+        return new_file
 
 
 @app.route('/', methods=['GET', 'POST'])
@@ -66,22 +70,23 @@ def upload_file():
         time = request.form['time']
         cntime = timedict[time][0]
 
-        try:
-            ext = request_files.filename.rsplit('.', maxsplit=1)[1]
-        except IndexError as e:
-            ext = '无扩展名'
-            user_info = [username, cntime, temperature, ext, bjt_time_s]
-            print('无扩展名错误：【' + str(e) + '】' + str(user_info))
-            flash(['照片格式错误，请重新上传', '微信告诉统计员，您用什么软件编辑了照片？'], 'danger')
-            return redirect(request.url)
-        else:
-            if ext not in IMAGES:
-                user_info = [username, cntime, temperature, ext, bjt_time_s]
-                print('可能是非图片扩展名：' + str(user_info))
+        if request_files.filename:
+            try:
+                ext = request_files.filename.rsplit('.', maxsplit=1)[1]
+            except IndexError as e:
+                ext = '无扩展名'
+                user_info = [username, cntime, temperature, request_files.filename, bjt_time_s]
+                print('01 无扩展名错误：【' + str(e) + '】' + str(user_info))
                 flash(['照片格式错误，请重新上传', '微信告诉统计员，您用什么软件编辑了照片？'], 'danger')
-                return redirect(request.url)
+                return jsonify(code='no_ext')
+            else:
+                if ext.lower() not in IMAGES:
+                    user_info = [username, cntime, temperature, request_files.filename, bjt_time_s]
+                    print('02 可能是非图片扩展名：' + str(user_info))
+                    flash(['照片格式错误，请重新上传', '微信告诉统计员，您用什么软件编辑了照片？'], 'danger')
+                    return jsonify(code='not_img')
 
-        user_info = [username, cntime, temperature, ext, bjt_time_s]
+        user_info = [username, cntime, temperature, request_files.filename, bjt_time_s]
 
         # 开始写入csv文件
         listcsv_file = app.config['UPLOADED_PHOTOS_DEST'] + 'listcsv.csv'
@@ -99,9 +104,9 @@ def upload_file():
                 row_num = name_list.index(username)
                 # 找到username在哪行
             except ValueError as e:
-                print(e)
+                print('03 姓名填错：【' + str(e) + '】')
                 flash(['【 ' + username + ' 】请输入正确的姓名', ''], 'warning')
-                return redirect(request.url)
+                return jsonify(code='wrong_name')
             else:
                 logcsv.seek(0)
                 # 把seek指针复位到开头。index后，指针到了最后
@@ -114,33 +119,50 @@ def upload_file():
                 # logcsv.seek(cellseek)
                 logcsv.write(temperature_f)
 
-        #开始保存图片
-        subfolder = bjt_date + '/' + bjt_date + time + '/'
-        rename = username + '+五部+' + bjt_date + cntime
-        photo_path = photos.save(request_files, folder=subfolder, name=rename + '.')
-        # 函数scale_tag(old_file, new_path, size, tag)修改图片尺寸，加标签
-        old_file = app.config['UPLOADED_PHOTOS_DEST'] + photo_path
-        new_path = app.config['UPLOADED_PHOTOS_DEST'] + bjt_date + '/' + bjt_date_year + cntime
-        if os.path.getsize(old_file) <= 4096:
-            os.remove(old_file)
-            print('0字节图片：' + str(user_info))
-            flash(['！！！照片上传失败 ！！！', '>>> 请重新上传 <<<'], 'danger')
-            return redirect(request.url)
-        else:
-            tag = username + ' ' + bjt_date + cntime + ' ' + temperature + '℃'
-            new_file = scale_tag(old_file, new_path, 1024, tag)
+        if request_files.filename:
+            #开始保存图片
+            subfolder = bjt_date + '/' + bjt_date + time + '/'
+            rename = username + '+五部+' + bjt_date + cntime
+            photo_path = photos.save(request_files, folder=subfolder, name=rename + '.')
+            # 函数scale_tag(old_file, new_path, size, tag)修改图片尺寸，加标签
+            old_file = app.config['UPLOADED_PHOTOS_DEST'] + photo_path
+            new_path = app.config['UPLOADED_PHOTOS_DEST'] + bjt_date + '/' + bjt_date_year + cntime
+            if os.path.getsize(old_file) <= 4096:
+                os.remove(old_file)
+                print('04 0字节图片：' + str(user_info))
+                flash(['！！！照片上传失败 ！！！', '>>> 请重新上传 <<<'], 'danger')
+                return jsonify(code='0bit')
+            else:
+                tag = username + ' ' + bjt_date + cntime + ' ' + temperature + '℃'
+                new_file = scale_tag(old_file, new_path, 1024, tag)
+                if new_file == 'fake_img':
+                    os.remove(old_file)
+                    print('05 图片扩展名，但文件不是图片：【' + new_file + '】' + str(user_info))
+                    flash(['照片格式错误，请重新上传', '微信告诉统计员，您用什么软件编辑了照片？'], 'danger')
+                    return jsonify(code='fake_img')
 
-        file_url = '_uploads/' + new_file
-        print(user_info)
-        status_message = ['success', '上报成功', user_info, file_url]
-        session['status_message'] = status_message
-        return redirect(url_for('show_images'))
+            file_url = '_uploads/' + new_file
+            print('06 正常上报图片：' + str(user_info))
+            status_message = ['success', '上报成功', user_info, file_url]
+            # session['status_message'] = status_message
+            return jsonify(status_message = status_message, code = 'is_photo')
+        else:
+            print('07 没上报图片：' + str(user_info))
+            status_message = ['success', '上报成功', user_info, '']
+            return jsonify(status_message = status_message, code = 'no_photo')
 
     return render_template('heat.html', time_message=bjt_date_y + ' ' + bjt_time)
 
-@app.route('/show')
-def show_images():
-    return render_template('show.html')
+# Error list:
+# 01 无扩展名错误
+# 02 可能是非图片扩展名
+# 03 姓名填错
+# 04 0字节图片
+# 05 图片扩展名，但文件不是图片
+# 06 正常上报图片
+# 07 没上报图片
+
+
 
 # 通过listx.html筛选csv、img文件用open.html打开，其他文件下载，目录进入
 @app.route('/listx/')
@@ -168,7 +190,7 @@ def list_file():
             file_url = '/listx/' + file
         url_list.append([url_filter, file_url, file])
     url_list.sort(reverse = True)
-    element_num = [folder_num, file_num]
+    element_num = ['首页', folder_num, file_num]
     return render_template('listx.html', url_list=url_list, abs_path=app.config['UPLOADED_PHOTOS_DEST'], element_num = element_num)
 
 @app.route('/listx/<path:path_name>')
@@ -197,7 +219,7 @@ def subpath_file(path_name):
             file_url = '/listx/' + escape(path_name) + '/' + file
         url_list.append([url_filter, file_url, file])
     url_list.sort()
-    element_num = [folder_num, file_num]
+    element_num = [path_name, folder_num, file_num]
     return render_template('listx.html', url_list=url_list, abs_path=abs_path, element_num = element_num)
 
 # 在浏览器页面展示csv和图片文件。
